feat: 개인 결제금액 코드 개선
