+++ src/modules/dutch/components/Calculate/index.ts
@@ -1,1 +1,2 @@
+export * from './PersonalCalculate';
 export * from './TotalPayment';
